Explicit double->float conversion
We were multiplying/dividing floats.

+++ PonykartParsers++/Misc/ogreExtensions.cpp
@@ -24,13 +24,13 @@ namespace Extensions
 	Vector3 degreeVectorToRadianVector(const Ogre::Vector3& vec)
 	{
 		const double twopi = 6.28318530717958647692; ///< 2*PI (double precision)
-		return Vector3(vec.x*twopi/360.0, vec.y*twopi/360.0, vec.z*twopi/360.0);
+		return Vector3((float)vec.x*twopi/360.0, (float)vec.y*twopi/360.0, (float)vec.z*twopi/360.0);
 	}
 
 	Vector3 radianVectorToDegreeVector(const Ogre::Vector3& vec)
 	{
 		const double twopi = 6.28318530717958647692; ///< 2*PI (double precision)
-		return Vector3(vec.x*360.0/twopi, vec.y*360.0/twopi, vec.z*360.0/twopi);
+		return Vector3((float)vec.x*360.0/twopi, (float)vec.y*360.0/twopi, (float)vec.z*360.0/twopi);
 	}
 
 	Quaternion degreeVectorToGlobalQuaternion(const Vector3& vec)
